moved business logic from controller for delete ingredient

+++ server/src/main/java/ua/kiev/naiv/drinkit/cocktail/service/impl/IngredientServiceImpl.java
@@ -5,11 +5,15 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import ua.kiev.naiv.drinkit.cocktail.exception.RecipesFoundException;
+import ua.kiev.naiv.drinkit.cocktail.exception.RecordNotFoundException;
 import ua.kiev.naiv.drinkit.cocktail.persistence.model.Ingredient;
+import ua.kiev.naiv.drinkit.cocktail.persistence.model.IngredientWithQuantity;
 import ua.kiev.naiv.drinkit.cocktail.persistence.repository.IngredientRepository;
 import ua.kiev.naiv.drinkit.cocktail.service.IngredientService;
 
 import java.util.List;
+import java.util.Set;
 
 @Service
 @Transactional
@@ -49,7 +53,16 @@ public class IngredientServiceImpl implements IngredientService {
     }
 
     @Override
-    public void delete(int id) {
-        ingredientRepository.delete(id);
+    public void delete(int id) throws RecipesFoundException {
+        Ingredient ingredient = ingredientRepository.findOne(id);
+        if (ingredient == null) {
+            throw new RecordNotFoundException("Ingredient not found : " + id);
+        }
+        Set<IngredientWithQuantity> cocktailIngredients = ingredient.getCocktailIngredients();
+        if (cocktailIngredients.isEmpty()) {
+            ingredientRepository.delete(id);
+        } else {
+            throw new RecipesFoundException(cocktailIngredients.size());
+        }
     }
 }
